exponent std::uint32_t -> std::int32_t

+++ src/support/fpbits.hh
@@ -11,7 +11,9 @@ class FpBits {
   FpBits(float x = 0.f) noexcept : bits_(bitCast<std::uint32_t>(x)) {}
   FpBits(std::uint32_t sig) noexcept : FpBits() { setSig(sig); }
   auto expBits() const noexcept { return bits_ & kExpMask; }
-  auto expValue() const noexcept { return (expBits() >> kSigLen) - kExpBias; }
+  auto expValue() const noexcept {
+    return static_cast<std::int32_t>(expBits() >> kSigLen) - kExpBias;
+  }
   auto sig() const noexcept { return bits_ & kSigMask; }
 
   auto getValue() const noexcept { return bitCast<float>(bits_); }
@@ -33,7 +35,7 @@ class FpBits {
   static constexpr std::uint32_t kExpMask = ((1 << kExpLen) - 1) << kSigLen;
   static constexpr std::uint32_t kSigMask = (1 << kSigLen) - 1;
 
-  static constexpr std::uint32_t kExpBias = 0x7f;
+  static constexpr std::int32_t kExpBias = 0x7f;
 
  private:
   std::uint32_t bits_;
